pyproject.toml
[project]
name = "telnet"
version = "0.1.0"
description = "TelNet - Teleconnection-based Neural Network for Seasonal Precipitation Forecasting"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cartopy>=0.22.0",
    "cdsapi>=0.7.0",
    "dask>=2024.1.0",
    "geopandas>=0.14.0",
    "h5netcdf>=1.3.0",
    "h5py>=3.10.0",
    "matplotlib>=3.8.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    "shapely>=2.0.0",
    "torch>=2.1.0",
    "tqdm>=4.66.0",
    "typer>=0.9.0",
    "xarray>=2024.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
telnet = "telnet.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["telnet"]
